+++ GW2EI.Executables/GW2EI.Applications/GW2EIParserAvalonia/Inspector/Models/EventModel.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using GW2EIEvtcParser.ParsedData;
+using GW2EIParserAvalonia.Services;
 
 namespace GW2EIParserAvalonia.Models;
